Remove keyboardfocus and field containers and make them separate packages
