+++ packages/selection/README.md
@@ -9,84 +9,9 @@ This package includes containers relating to selection in the
 npm install @zendeskgarden/container-selection
 ```
 
-## Containers & Hooks
-
-This package contains the following render-prop containers and their corresponding hooks
-
-- FieldContainer / useField
-- KeyboardFocusContainer / useKeyboardFocus
-- SelectionContainer / useSelection
-
 ## Usage
 
-For live examples check out our [storybook](https://zendeskgarden.github.io/react-containers).
-
-### FieldContainer
-
-FieldContainer is a render-prop wrapper for the `useField` hook which will add the for/id
-mapping and `aria-labeledby` mapping when a hint is applied.
-
-```jsx static
-import { FieldContainer } from '@zendeskgarden/container-selection';
-
-<FieldContainer>
-  {({ getLabelProps, getInputProps, getHintProps }) => (
-    <>
-      <label {...getLabelProps()}>Accessible Native Input</label>
-      <p {...getHintProps()}>Optional Hint</p>
-      <input {...getInputProps()} />
-    </>
-  )}
-</FieldContainer>;
-```
-
-### useField
-
-```jsx static
-import { FieldContainer } from '@zendeskgarden/container-selection';
-
-const Field = () => {
-  const { getLabelProps, getInputProps, getHintProps } = useField();
-
-  return (
-    <>
-      <label {...getLabelProps()}>Accessible Native Input</label>
-      <p {...getHintProps()}>Optional Hint</p>
-      <input {...getInputProps()} />
-    </>
-  );
-};
-```
-
-### KeyboardFocusContainer
-
-KeyboardFocusContainer is a render-prop wrapper for the `useKeyboardFocus` hook which allows
-you to distinguish between mouse and keyboard focus. Garden uses this in react-components to
-know when to add the focus ring.
-
-```jsx static
-import { KeyoardFocusContainer } from '@zendeskgarden/container-selection';
-
-<KeyboardFocusContainer>
-  {({ keyboardFocused, getFocusProps }) => (
-    <div {...getFocusProps()}>{keyboardFocused ? 'Keyboard focused!' : 'Not keyboard focused'}</div>
-  )}
-</KeyboardFocusContainer>;
-```
-
-### useKeyboardFocus
-
-```jsx static
-import { useKeyboardFocus } from '@zendeskgarden/container-selection';
-
-const KeyboardFocus = () => {
-  const { getFocusProps, keyboardFocused } = useKeuyboardFocus();
-
-  return (
-    <div {...getFocusProps()}>{keyboardFocused ? 'Keyboard focused!' : 'Not keyboard focused'}</div>
-  );
-};
-```
+For live examples check out our [storybook](https://zendeskgarden.github.io/react-containers?selectedKind=Selection%20Containers).
 
 ### SelectionContainer
