--- packages/selection/src/containers/FieldContainer.js
@@ -1,21 +1,0 @@
-/**
- * Copyright Zendesk, Inc.
- *
- * Use of this source code is governed under the Apache License, Version 2.0
- * found at http://www.apache.org/licenses/LICENSE-2.0.
- */
-
-import PropTypes from 'prop-types';
-
-import { useField } from '../hooks/useField';
-
-export function FieldContainer({ children, render = children, id }) {
-  return render(useField(id));
-}
-
-/** TODO: Update prop-types */
-FieldContainer.propTypes = {
-  children: PropTypes.func,
-  render: PropTypes.func,
-  id: PropTypes.string
-};
