+++ packages/selection/src/index.spec.js
@@ -8,7 +8,6 @@
 import { getExports } from '@zendeskgarden/react-testing';
 import * as rootIndex from './';
 import * as locale from './components/LocaleProvider';
-import * as IdManager from './utils/IdManager';
 
 describe('Index', () => {
   it('exports all components and utilities', async () => {
@@ -22,8 +21,8 @@ describe('Index', () => {
               .split('/')
               .pop()
           )
-          .filter(file => !/(ACTIONS|DIRECTIONS|getControlledValue|IdManager)/u.test(file))
-          .concat(Object.keys(locale).filter(l => l === 'LocaleContext'), Object.keys(IdManager))
+          .filter(file => !/(ACTIONS|DIRECTIONS|getControlledValue)/u.test(file))
+          .concat(Object.keys(locale).filter(l => l === 'LocaleContext'))
           .sort();
       }
     });
