+++ packages/selection/src/useSelection.js
@@ -7,12 +7,12 @@
 
 import { useEffect, useContext, useReducer } from 'react';
 
-import { LocaleContext } from '../components/LocaleProvider';
-import { composeEventHandlers } from '../utils/composeEventHandlers';
-import { getControlledValue } from '../utils/getControlledValue';
-import { KEY_CODES } from '../utils/KEY_CODES';
-import { DIRECTIONS } from '../utils/DIRECTIONS';
-import { ACTIONS } from '../utils/ACTIONS';
+import { LocaleContext } from './components/LocaleProvider';
+import { composeEventHandlers } from './utils/composeEventHandlers';
+import { getControlledValue } from './utils/getControlledValue';
+import { KEY_CODES } from './utils/KEY_CODES';
+import { DIRECTIONS } from './utils/DIRECTIONS';
+import { ACTIONS } from './utils/ACTIONS';
 
 /* eslint-disable indent */
 function stateReducer(state, action, { focusedItem, selectedItem, onFocus, onSelect }) {
